refactor(punishments): share the punishment message helpers between both commands
